+++ tests/TestCase.php
@@ -2,9 +2,10 @@
 
 namespace Bavix\WalletSwap\Test;
 
-use Bavix\WalletSwap\WalletSwapServiceProvider;
+use Bavix\WalletSwap\RateServiceProvider;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 use Illuminate\Foundation\Application;
+use Swap\Laravel\SwapServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
@@ -15,7 +16,10 @@ class TestCase extends OrchestraTestCase
      */
     protected function getPackageProviders($app): array
     {
-        return [WalletSwapServiceProvider::class];
+        return [
+            SwapServiceProvider::class,
+            RateServiceProvider::class,
+        ];
     }
 
     /**
@@ -26,11 +30,13 @@ class TestCase extends OrchestraTestCase
      */
     protected function getEnvironmentSetUp($app): void
     {
-        $app['config']->set('database.default', 'testbench');
-        $app['config']->set('database.connections.testbench', [
-            'driver' => 'sqlite',
-            'database' => ':memory:',
-            'prefix' => '',
+        $app['config']->set('swap.services', [
+            'national_bank_of_romania' => true,
+            'central_bank_of_republic_turkey' => true,
+            'central_bank_of_czech_republic' => true,
+            'russian_central_bank' => true,
+            'webservicex' => true,
+            'cryptonator' => true,
         ]);
     }